Implement new equipment registration form and enhance equipment listing with additional details
