+++ client/src/app/equipament/page.tsx
@@ -9,29 +9,63 @@ type Equipament = {
   name: string;
   patrimonio: string;
   tag: string;
+  serie: string;
   modelo: string;
   fabricante: string;
   status: string;
+  locationId: number;
+  equipamentTypeId: number;
+};
+
+type Location = {
+  id: number;
+  block: string;
+  room: string;
+};
+
+type EquipamentType = {
+  id: number;
+  name: string;
 };
 
 export default function EquipamentListPage() {
   const { fetchWithAuth } = useFetch("Listagem de Equipamentos");
-  const [equipaments, setEquipaments] = useState<Equipament[]>([]);
   const router = useRouter();
 
-  const fetchEquipaments = async () => {
-    const result = await fetchWithAuth("/equipament", {
-      method: "GET",
-    });
+  const [equipaments, setEquipaments] = useState<Equipament[]>([]);
+  const [locations, setLocations] = useState<Location[]>([]);
+  const [equipamentTypes, setEquipamentTypes] = useState<EquipamentType[]>([]);
 
-    if (result?.status === 200) {
-      setEquipaments(result.data);
-    }
+  // Buscar dados de equipamentos, locais e tipos
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const [equipamentRes, locRes, typeRes] = await Promise.all([
+          fetchWithAuth("/equipament", { method: "GET" }),
+          fetchWithAuth("/location", { method: "GET" }),
+          fetchWithAuth("/equipament-type", { method: "GET" }),
+        ]);
+
+        if (equipamentRes?.status === 200) setEquipaments(equipamentRes.data);
+        if (locRes?.status === 200) setLocations(locRes.data);
+        if (typeRes?.status === 200) setEquipamentTypes(typeRes.data);
+      } catch (err) {
+        console.error("Erro ao buscar dados:", err);
+      }
+    };
+
+    fetchData();
+  }, []);
+
+  const getLocationName = (id: number) => {
+    const loc = locations.find((l) => l.id === id);
+    return loc ? `${loc.block} / Sala ${loc.room}` : "-";
   };
 
-  useEffect(() => {
-    fetchEquipaments();
-  }, []);
+  const getTypeName = (id: number) => {
+    const type = equipamentTypes.find((t) => t.id === id);
+    return type?.name || "-";
+  };
 
   return (
     <main className="min-h-screen p-10 bg-gray-100">
@@ -44,30 +78,31 @@ export default function EquipamentListPage() {
               <th className="py-3 px-4 text-left">Nome</th>
               <th className="py-3 px-4 text-left">Patrimônio</th>
               <th className="py-3 px-4 text-left">Tag</th>
+              <th className="py-3 px-4 text-left">Nº de Série</th>
               <th className="py-3 px-4 text-left">Modelo</th>
               <th className="py-3 px-4 text-left">Fabricante</th>
               <th className="py-3 px-4 text-left">Status</th>
+              <th className="py-3 px-4 text-left">Local</th>
+              <th className="py-3 px-4 text-left">Tipo Equipamento</th>
               <th className="py-3 px-4 text-left">Ações</th>
             </tr>
           </thead>
           <tbody>
-            {equipaments.map((equipament) => (
-              <tr
-                key={equipament.id}
-                className="border-t hover:bg-gray-50 transition"
-              >
-                <td className="py-3 px-4">{equipament.name}</td>
-                <td className="py-3 px-4">{equipament.patrimonio}</td>
-                <td className="py-3 px-4">{equipament.tag}</td>
-                <td className="py-3 px-4">{equipament.modelo}</td>
-                <td className="py-3 px-4">{equipament.fabricante}</td>
-                <td className="py-3 px-4">{equipament.status}</td>
+            {equipaments.map((equip) => (
+              <tr key={equip.id} className="border-t hover:bg-gray-50 transition">
+                <td className="py-3 px-4">{equip.name}</td>
+                <td className="py-3 px-4">{equip.patrimonio}</td>
+                <td className="py-3 px-4">{equip.tag}</td>
+                <td className="py-3 px-4">{equip.serie}</td>
+                <td className="py-3 px-4">{equip.modelo}</td>
+                <td className="py-3 px-4">{equip.fabricante}</td>
+                <td className="py-3 px-4">{equip.status}</td>
+                <td className="py-3 px-4">{getLocationName(equip.locationId)}</td>
+                <td className="py-3 px-4">{getTypeName(equip.equipamentTypeId)}</td>
                 <td className="py-3 px-4">
                   <button
                     className="text-blue-600 hover:underline"
-                    onClick={() =>
-                      router.push(`/equipament/${equipament.id}`)
-                    }
+                    onClick={() => router.push(`/equipament/${equip.id}`)}
                   >
                     Ver detalhes
                   </button>
@@ -76,7 +111,7 @@ export default function EquipamentListPage() {
             ))}
             {equipaments.length === 0 && (
               <tr>
-                <td colSpan={7} className="text-center py-6 text-gray-500">
+                <td colSpan={10} className="text-center py-6 text-gray-500">
                   Nenhum equipamento cadastrado.
                 </td>
               </tr>
