+++ client/src/components/ButtonByRole.tsx
@@ -37,7 +37,7 @@ export default function ButtonByRole({ role }: { role: ROLE }) {
     return (
         <button
             onClick={() => setClick(true)}
-            className="p-4 border border-gray-300 rounded-md hover:bg-green-100"
+            className="text-sm px-3 py-1 border border-gray-300 rounded-md  hover:bg-gray-200 transition"
         >
             Consultar {role}
         </button>
